pyproject.toml
[tool.poetry]
name = "LearnPythonAPI"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
allure-pytest = "^2.9.45"
requests = "^2.28.0"
pytest = "^7.1.2"
lxml = "^4.9.0"
yarl = "^1.7.2"
pydantic = "^1.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
